refactor BlueprintsPageLayout -> ModsPageLayout
