+++ src/extensionConsole/pages/mods/ModsPageLayout.test.tsx
@@ -17,7 +17,7 @@
 
 import React from "react";
 import { render } from "@/extensionConsole/testHelpers";
-import BlueprintsPageLayout from "@/extensionConsole/pages/mods/BlueprintsPageLayout";
+import ModsPageLayout from "@/extensionConsole/pages/mods/ModsPageLayout";
 import { type Mod } from "@/mods/modTypes";
 import { waitForEffect } from "@/testUtils/testHelpers";
 import { act, screen } from "@testing-library/react";
@@ -43,7 +43,7 @@ jest.mock("@/recipes/recipesHooks", () => ({
 
 const installables: Mod[] = [];
 
-describe("BlueprintsPageLayout", () => {
+describe("ModsPageLayout", () => {
   const { env } = process;
 
   beforeEach(() => {
@@ -60,9 +60,7 @@ describe("BlueprintsPageLayout", () => {
   });
 
   test("renders", async () => {
-    const rendered = render(
-      <BlueprintsPageLayout installables={installables} />
-    );
+    const rendered = render(<ModsPageLayout mods={installables} />);
     await waitForEffect();
     expect(rendered.asFragment()).toMatchSnapshot();
   });
@@ -72,7 +70,7 @@ describe("BlueprintsPageLayout", () => {
 
     const deferred = onDeferredGet("/api/onboarding/starter-blueprints/");
 
-    render(<BlueprintsPageLayout installables={installables} />);
+    render(<ModsPageLayout mods={installables} />);
     await waitForEffect();
     expect(
       screen.queryByText("Welcome to the PixieBrix Extension Console")
@@ -89,7 +87,7 @@ describe("BlueprintsPageLayout", () => {
   });
 
   test("get started tab is active by default", async () => {
-    render(<BlueprintsPageLayout installables={installables} />);
+    render(<ModsPageLayout mods={installables} />);
     await waitForEffect();
     expect(
       screen.queryByText("Welcome to the PixieBrix Extension Console")
@@ -106,7 +104,7 @@ describe("BlueprintsPageLayout", () => {
         organization: userOrganizationFactory(),
       })
     );
-    render(<BlueprintsPageLayout installables={installables} />);
+    render(<ModsPageLayout mods={installables} />);
     await waitForEffect();
     expect(
       screen.queryByText("Welcome to the PixieBrix Extension Console")
@@ -117,7 +115,7 @@ describe("BlueprintsPageLayout", () => {
   test("shows the bot games tab", async () => {
     mockCachedUser();
 
-    render(<BlueprintsPageLayout installables={installables} />, {
+    render(<ModsPageLayout mods={installables} />, {
       setupRedux(dispatch) {
         dispatch(
           authSlice.actions.setAuth(
@@ -137,11 +135,11 @@ describe("BlueprintsPageLayout", () => {
   test("doesn't flash get started tab while loading the bot games tab", async () => {
     mockLoadingUser();
 
-    render(<BlueprintsPageLayout installables={installables} />);
+    render(<ModsPageLayout mods={installables} />);
     await waitForEffect();
     expect(screen.queryByText("Get Started")).toBeNull();
 
-    render(<BlueprintsPageLayout installables={installables} />, {
+    render(<ModsPageLayout mods={installables} />, {
       setupRedux(dispatch) {
         dispatch(
           authSlice.actions.setAuth(
@@ -159,7 +157,7 @@ describe("BlueprintsPageLayout", () => {
   });
 
   test("bot games tab is active by default", async () => {
-    render(<BlueprintsPageLayout installables={installables} />, {
+    render(<ModsPageLayout mods={installables} />, {
       setupRedux(dispatch) {
         dispatch(
           authSlice.actions.setAuth(
@@ -178,7 +176,7 @@ describe("BlueprintsPageLayout", () => {
 
   test("search query heading renders", async () => {
     const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
-    render(<BlueprintsPageLayout installables={installables} />);
+    render(<ModsPageLayout mods={installables} />);
 
     await waitForEffect();
 
@@ -207,7 +205,7 @@ describe("BlueprintsPageLayout", () => {
 describe("Serializable Data Test", () => {
   test("Pushes unserializable data to redux", async () => {
     const spy = jest.spyOn(console, "error");
-    render(<BlueprintsPageLayout installables={installables} />, {
+    render(<ModsPageLayout mods={installables} />, {
       setupRedux(dispatch) {
         dispatch(
           blueprintsSlice.actions.setSearchQuery(
